% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_scores.R
\name{calculate_scores}
\alias{calculate_scores}
\title{calculate_scores}
\usage{
calculate_scores(feature_df)
}
\arguments{
\item{feature_df}{A data frame containing features for each position.}
}
\value{
The function returns a data frame with calculated scores for each position.
}
\description{
Calculates various scores for each position in the feature data frame.
}
