% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_config.R
\name{check_config}
\alias{check_config}
\title{check_config}
\usage{
check_config(config_file = "")
}
\arguments{
\item{config_file}{A character string specifying the path to the configuration file.}
}
\description{
Checks if a configuration file exists and loads the configuration parameters. If the file does not exist, it sets default values for the configuration parameters.
}
\examples{
check_config()
}
