% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findExecutable.R
\name{.findExecutable}
\alias{.findExecutable}
\title{findExecutable}
\usage{
.findExecutable(exe, interactive = TRUE)
}
\arguments{
\item{exe}{A character string specifying the name of the executable to be found.}

\item{interactive}{A logical value indicating whether to stop with an error message if the executable is not found.}
}
\value{
The function returns the path to the executable or an empty character vector if the executable is not found and interactive is set to FALSE.
}
\description{
Find the location of an executable in the system's PATH environment variable.
}
\examples{
# Find the location of the "muscle" executable
#muscle_path <- .findExecutable("muscle")
#muscle_path
}
