% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dssp_command.R
\name{dssp_command}
\alias{dssp_command}
\title{dssp_command}
\usage{
dssp_command(pdb_file)
}
\arguments{
\item{pdb_file}{A character vector specifying the path to the input PDB file.}
}
\value{
The function returns the name of the output DSSP file.
}
\description{
Runs the DSSP command with mkdssp to process a PDB file and generates a DSSP file.
}
\examples{
# Generate the DSSP file from the input PDB file "protein.pdb"
#dssp_command("protein.pdb")

}
