% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ftp_download.R
\name{ftp_download}
\alias{ftp_download}
\title{ftp_download}
\usage{
ftp_download(link, seq_type = "pep")
}
\arguments{
\item{link}{A character vector specifying the species link to download.}

\item{seq_type}{A character vector specifying the sequence type (default is "pep").}
}
\value{
The downloaded file will be saved to the specified destination folder.
}
\description{
Downloads the Ensembl CDS file for a specified species from the NCBI FTP site.
}
\examples{
# Download the CDS file for Mus musculus (protein sequences)
#ftp_download("Mus_musculus/")
}
