% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iupredAnchor.R
\name{iupredAnchor}
\alias{iupredAnchor}
\title{iupredAnchor}
\usage{
iupredAnchor(uniprotAccession, proteinName = NA)
}
\arguments{
\item{uniprotAccession}{A character string specifying the UniProt accession number of the protein.}

\item{proteinName}{A character string specifying the name of the protein (optional).}
}
\value{
The function returns a data frame containing the predicted intrinsic disorder and anchor residues at each position in the protein sequence. If \code{plotResults = TRUE}, it returns a ggplot object showing the prediction plot. If \code{plotResults = FALSE}, it returns the data frame only.
}
\description{
Predicts intrinsic disorder and anchor residues in a protein using the IUPred2A web service.
}
