% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/muscle.R
\name{muscle}
\alias{muscle}
\title{muscle}
\usage{
muscle(x)
}
\arguments{
\item{x}{A set of biological sequences (RNAStringSet, DNAStringSet, or AAStringSet).}
}
\value{
The function returns the multiple sequence alignment result as an RNAStringSet, DNAStringSet, or AAStringSet, based on the input sequence type.
}
\description{
Perform multiple sequence alignment using the MUSCLE program.
}
\examples{
# Create a set of DNA sequences
#dna_sequences <- DNAStringSet(c("ATCGATCG", "AT-AT-AT", "GCG-GCG"))
# Perform multiple sequence alignment
#alignment <- muscle(dna_sequences)

}
