% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_dssp.R
\name{parse_dssp}
\alias{parse_dssp}
\title{parse_dssp}
\usage{
parse_dssp(file, keepfiles = TRUE)
}
\arguments{
\item{file}{A character vector specifying the path to the input DSSP file.}

\item{keepfiles}{A logical value indicating whether to keep the parsed data in an Rda file (default is FALSE).}
}
\value{
The function returns a data frame containing the parsed DSSP data.
}
\description{
Parses a DSSP file and returns the data in a data frame.
}
\examples{
#parse_dssp("protein.dssp", keepfiles = TRUE)
}
