% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protein_blast.R
\name{protein_blast}
\alias{protein_blast}
\title{protein_blast}
\usage{
protein_blast(seqs, blast_file, blast_type = "blastp")
}
\arguments{
\item{seqs}{A character vector of input sequences to be used in the BLAST search.}

\item{blast_file}{A character vector specifying the path to the BLAST database file.}

\item{blast_type}{A character vector specifying the type of BLAST (default is "blastp").}
}
\value{
The function returns the BLAST results as a list, including the matches and information about the BLAST database used.
}
\description{
Performs protein BLAST using the specified BLAST database file and predicts matches for input sequences.
}
\examples{
# Perform protein BLAST using the "database.fasta" database file on input sequences
#blast_results <- protein_blast(seqs, blast_file = "database.fasta", blast_type = "blastp")

}
