% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_uniprot.R
\name{query_uniProt}
\alias{query_uniProt}
\title{query_uniProt}
\usage{
query_uniProt(
  query = character(0L),
  fields = c("accession", "id"),
  collapse = " OR"
)
}
\arguments{
\item{query}{A character vector specifying the query terms to search for in UniProt.}

\item{fields}{A character vector specifying the fields to include in the query results (default is c("accession", "id")).}

\item{collapse}{A character vector specifying the separator for combining multiple query terms (default is " OR").}
}
\value{
A data frame containing the results of the UniProt query.
}
\description{
Queries the UniProt API and retrieves information based on the provided query and fields.
}
\examples{
# Query UniProt for information about proteins with accessions "P04114" and "P15056"
query_uniProt(query = c("P04114", "P15056"))

}
