% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shannon_entropy.R
\name{shannon_entropy}
\alias{shannon_entropy}
\title{shannon_entropy}
\usage{
shannon_entropy(seq, nogap = FALSE)
}
\arguments{
\item{seq}{A vector representing the sequence for which the entropy is to be calculated.}

\item{nogap}{A logical value indicating whether to exclude gaps (default is FALSE).}
}
\value{
The function returns the Shannon entropy of the input sequence.
}
\description{
Calculates the Shannon entropy of a sequence.
}
\examples{
# Calculate the Shannon entropy of a sequence
entropy <- shannon_entropy(c("A", "A", "T", "C", "G"))

}
