% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shannon_reshape.R
\name{shannon_reshape}
\alias{shannon_reshape}
\title{shannon_reshape}
\usage{
shannon_reshape(msa, query)
}
\arguments{
\item{msa}{A multiple sequence alignment object (e.g., from the `msa` package).}

\item{query}{The index or name of the query sequence in the MSA.}
}
\value{
The function returns a data frame containing the positions, amino acids, and their corresponding Shannon entropy values for the query sequence.
}
\description{
Calculates Shannon entropy for each position of the query sequence in the multiple sequence alignment (MSA).
}
