% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ss_convert.R
\name{ss_convert}
\alias{ss_convert}
\title{ss_convert}
\usage{
ss_convert(.x)
}
\arguments{
\item{.x}{A character vector representing the Secondary Structure Symbol to be converted.}
}
\value{
The function returns the numeric value corresponding to the Secondary Structure Symbol as defined in the SS_key.
}
\description{
Converts a Secondary Structure Symbol to its corresponding numeric value using the SS_key.
}
\examples{
# Convert Secondary Structure Symbols to numeric values
ss_convert("H")  # Output: 1
ss_convert("E")  # Output: 2

}
