% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dssp_command.R
\name{dssp_command}
\alias{dssp_command}
\title{dssp_command}
\usage{
dssp_command(cif_file)
}
\arguments{
\item{cif_file}{A character vector specifying the path to the input mmCIF file.}
}
\value{
The function returns the name of the output DSSP file.
}
\description{
Runs the DSSP command with mkdssp to process a mmCIF file and generates a DSSP file.
}
\examples{
# Generate the DSSP file from the input mmCIF file "protein.cif"
#dssp_command("protein.cif")

}
