% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_alphafold.R
\name{fetch_alphafold}
\alias{fetch_alphafold}
\title{fetch_alphafold}
\usage{
fetch_alphafold(protein_id)
}
\arguments{
\item{protein_id}{A character vector specifying the protein ID for which the prediction is to be fetched.}
}
\value{
If successful, the function returns the path to the downloaded Alphafold2 mmCIF file. If the file already
exists in the specified directory, a message is displayed, and the existing file path is returned. If the download
fails, an error message is displayed, and the function returns NULL.
}
\description{
Downloads the Alphafold2 protein structure prediction in mmCIF format based on the provided protein ID.
}
\examples{
# Download the Alphafold2 mmCIF for a protein with ID "P57102"
fetch_alphafold("P57102")

}
